integrating blocks
